requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asteval==1.0.8
astunparse==1.6.3
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
blinker==1.9.0
boto3==1.42.69
botocore==1.42.69
CacheControl==0.14.4
cached-property==2.0.1
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
chardet==7.1.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.5
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
deepdiff==7.0.1
Deprecated==1.3.1
deptry==0.24.0
diff-match-patch==20241021
Django==5.2.12
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-import-export==4.4.0
django-prometheus==2.4.1
django-redis==6.0.0
django-request-logging==0.7.5
django-storages==1.14.6
django-timezone-field==7.2.1
django-unfold==0.85.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-yasg==1.21.15
et_xmlfile==2.0.0
execnet==2.1.2
firebase_admin==7.2.0
flexcache==0.3
flexparser==0.4
flower==2.0.1
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
ijson==3.5.0
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
init-data-py==0.2.7
isodate==0.7.2
jmespath==1.1.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
loguru==0.7.3
lxml==6.0.2
morefs==0.2.2
msgpack==1.1.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==26.0
peewee==4.0.2
phonenumbers==9.0.26
pillow==12.1.1
Pint==0.25.2
platformdirs==4.9.4
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
PyPrind==2.11.3
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
pytest-xdist==3.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-telegram-bot==22.7
pytz==2026.1.post1
pyxlsb==1.0.10
PyYAML==6.0.3
RapidFuzz==3.14.3
rdflib==7.6.0
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rfc3986==2.0.0
rpds-py==0.30.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
structlog==25.5.0
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
unicodecsv==0.14.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.12.0
wrapt==2.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
zope.interface==8.2
